+++ examples/todo/todo/database.py
@@ -1,0 +1,3 @@
+# Just a mock database, for demonstration purposes
+todo_id_sequence = 0
+todo_database = {}
