+++ examples/todo/wsgi.py
@@ -1,0 +1,3 @@
+from todo.app_init import create_app
+
+app = create_app()
